# vulkan: subgroup size tuning