--- a/ggml/src/ggml-vulkan/ggml-vulkan.cpp
+++ b/ggml/src/ggml-vulkan/ggml-vulkan.cpp
@@ -1507,12 +1507,12 @@ static bool ggml_vk_matmul_shmem_support(const vk_device& device, const std::vec
 }
 
 struct GpuPipelineConfig {
-    // List of all aliases for a given GPU.
-    // For example, this can include names like "NAVI10", "RX 5700", etc.
-    std::vector<std::string> device_names;
+    // GPU architecture identifier.
+    // Example: vk_device_architecture::AMD_GCN
+    vk_device_architecture arch;
 
     // Mapping of pipeline names to their specific subgroup sizes.
-    // Example: {"soft_max_f32", 64}.
+    // Example: {"soft_max_f32", 64}
     std::unordered_map<std::string, uint32_t> pipelines;
 
     // Default subgroup size for this GPU.
@@ -1523,7 +1523,7 @@ struct GpuPipelineConfig {
 // Define configurations for different GPUs.
 static std::vector<GpuPipelineConfig> gpu_pipeline_configs = {
     {
-        {"NAVI10", "NAVI14", "RX 5700", "RX 5600", "RX 5500"},
+        vk_device_architecture::AMD_RDNA1,
         {
             {"soft_max_f32", 64}, {"soft_max_f32_wg512", 64},
             {"soft_max_f32_f16", 64}, {"soft_max_f32_f16_wg512", 64},
@@ -1533,16 +1533,14 @@ static std::vector<GpuPipelineConfig> gpu_pipeline_configs = {
     },
 };
 
-static uint32_t get_subgroup_size(const std::string &pipeline_name, const std::string &device_name) {
+static uint32_t get_subgroup_size(const std::string &pipeline_name, const vk_device_architecture &arch) {
     for (const auto &config : gpu_pipeline_configs) {
-        for (const auto &alias : config.device_names) {
-            if (device_name.find(alias) != std::string::npos) {
-                auto pipIt = config.pipelines.find(pipeline_name);
-                if (pipIt != config.pipelines.end() && pipIt->second != 0) {
-                    return pipIt->second;
-                }
-                return config.default_subgroup_size;
+        if (config.arch == arch) {
+            auto pipIt = config.pipelines.find(pipeline_name);
+            if (pipIt != config.pipelines.end() && pipIt->second != 0) {
+                return pipIt->second;
             }
+            return config.default_subgroup_size;
         }
     }
     // If no matching configuration is found, return 0.
@@ -1670,16 +1668,12 @@ static void ggml_vk_load_shaders(vk_device& device) {
         device->pipeline_matmul_id_f32 = std::make_shared<vk_matmul_pipeline_struct>();
     }
 
-    vk::PhysicalDeviceProperties2 props2;
-    device->physical_device.getProperties2(&props2);
-    std::string device_name = props2.properties.deviceName.data();
-
     std::vector<std::future<void>> compiles;
     auto const &ggml_vk_create_pipeline = [&](vk_device& device, vk_pipeline& pipeline, const std::string &name, size_t spv_size, const void* spv_data, const std::string &entrypoint,
                                               uint32_t parameter_count, uint32_t push_constant_size, std::array<uint32_t, 3> wg_denoms, const std::vector<uint32_t>& specialization_constants,
                                               uint32_t align, bool disable_robustness = false, bool require_full_subgroups = false, uint32_t required_subgroup_size = 0) {
 
-        required_subgroup_size = get_subgroup_size(name, device_name);
+        required_subgroup_size = get_subgroup_size(name, device->architecture);
 
         if (!pipeline) {
             pipeline = std::make_shared<vk_pipeline_struct>();
@@ -2890,7 +2884,8 @@ static void ggml_vk_print_gpu_info(size_t idx) {
     subgroup_props.pNext = &driver_props;
     physical_device.getProperties2(&props2);
 
-    uint32_t default_subgroup_size = get_subgroup_size("", props2.properties.deviceName.data());
+    vk_device_architecture arch = get_device_architecture(physical_device);
+    uint32_t default_subgroup_size = get_subgroup_size("", arch);
     const size_t subgroup_size = (default_subgroup_size != 0) ? default_subgroup_size : subgroup_props.subgroupSize;
 
     const bool uma = props2.properties.deviceType == vk::PhysicalDeviceType::eIntegratedGpu;