@0cc4m 2025-03-08
I tried doing device architecture detection in a better way in https://github.com/daniandtheweb/llama.cpp/pull/1, let me know what you think.