@0cc4m 2025-03-11
I'll give this another try on RDNA2 soon, I expect that you can add RDNA2 to the RDNA1 logic. Once that is done I think we can merge this.

Can you resolve the merge conflict?