@0cc4m 2025-03-11
I don't know what it is, but to figure out which op is failing I implemented the `GGML_VK_CHECK_RESULTS` thing. Try to compile with that, it should tell you which result is wrong.