@0cc4m 2025-03-11
> For some reason all mul_mat and mul_mat_id fail on RDNA3 when using a subgroup size of 32 (I was in a rush before so I wasn't able to run the test-backend-ops) so I'll just avoid doing changes on this arch for now until I find a proper tuning for this card.

That's a coopmat thing, I think. If you ignore coopmat shaders it's probably fine.