@0cc4m 2025-03-12
I checked and you can add RDNA2 to the list. It seems to prefer the same shader sizes. Afterwards we can merge this.