@daniandtheweb 2025-03-12
I created a common RDNA map to store the subgroup preferences since they mostly behave the same, however I've kept the option to specify different configurations just in case better tuning options are found for a specific architecture. I'll make more tests with other subgroup configurations but if I find something better I'll leave it for another pr.
If you think this looks good I think it's ready to be merged.