@daniandtheweb 2025-03-12
I disabled warp 32 for RDNA3 for now as it seems to behave worse that with warp 64 when using the matrix cores, the bigger performance hits being on MUL_MAT f32 f32 and f16 f32. Every other operation seems mostly unaffected but since this results in no performance improvement I'll leave it untouched for now.

```
MUL_MAT(type_a=f32,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]): 288.38 GFLOPS  243.03 GFLOPS  -45.35 GFLOPS
MUL_MAT(type_a=f16,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]): 517.14 GFLOPS  437.11 GFLOPS  -80.03 GFLOP
MUL_MAT(type_a=f32,type_b=f32,m=4096,n=2,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]): 568.25 GFLOPS  497.69 GFLOPS  -70.56 GFLOPS
MUL_MAT(type_a=f16,type_b=f32,m=4096,n=2,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]): 1000.0 GFLOPS  876.59 GFLOPS  -123.41 GFLOPS
```