@0cc4m 2025-03-12
Matrix cores only get used with coopmat, and that's only for matrix matrix multiplications. What you're looking at is just a regular shader. But interesting that it performs worse with subgroup size 32, I guess I should check RDNA2 and 1 again as well.