@daniandtheweb 2025-03-12
I'm currently running all the performance tests on both RDNA1 AN RDNA3 to see if a finer tuning of the subgroup sizes can improve the situation.