@daniandtheweb 2025-03-13
After running multiple tests I can say for certain that RDNA3 doesn't like subgroup size 32, at least on linux using amdgpu (I tried radv, amdvlk and vk_pro and all behave similarly), every single operation seems to get a small performance hit from using it. I'm not sure if this is an expected behavior or there's some issue in the amdgpu driver (I found other issues in other compute tasks so it's a possibility). For now I don't think it's a good idea changing the default subgroup for this architecture.

Regarding RDNA1 I managed to solve every single perfomance regression by setting most mul_mat_vec operations to use subgroup 64. All the performance gains from switching to subgroup size 32 are maintained while not having any regression on any operation (at least the ones in test-backend-ops). I'll push the changes soon.

If you can compare the results using warp 64 and warp 32 on RDNA2 I can tune the parameters that still need some adjustements if there's any need for it. Does the performance on RDNA2 improve at all with these changes or it behaves similarly to how I described RDNA3?