@0cc4m 2025-03-16
> Thanks for the results. GFX1013 is only the Playstation 5 GPU or is there any other card using that chip? Is the performance using llama-bench or stable-diffusion.cpp increased on any of those two cards using subgroup 32 or is there any performance regression in those real world use cases?

It's an AMD BC-250, based on the PS5 APU, yeah. Actual performance in llama-bench does increase with your RDNA1 tuning:

| model                          |       size |     params | backend    | ngl |          test |                  t/s Master | t/s PR | t/s subgroup 32 |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | -------------------: | -----------------: | ------------------: |
| llama 8B Q4_0                  |   4.33 GiB |     8.03 B | Vulkan     |  99 |         pp512 |        278.58 ± 0.27 |        313.68 ± 0.04 |        313.23 ± 0.05 |
| llama 8B Q4_0                  |   4.33 GiB |     8.03 B | Vulkan     |  99 |         tg128 |         45.41 ± 0.46 |         44.96 ± 0.02 |         45.43 ± 0.03 |

There might be a small regression in mul_mat_vec when you keep it at 64, but it's hard to tell.